+++ src/FlatValidator/FlatValidationResult.cs
@@ -6,18 +6,6 @@ namespace System.Validation;
 /// <summary>
 /// Define a validation error contaner
 /// </summary>
-[Serializable]
-public record class FlatValidationError(string PropertyName, string ErrorMessage)
-{
-    /// <summary>
-    /// Userfriendly representation of the error
-    /// </summary>
-    public override string ToString() => $"{PropertyName}: {ErrorMessage}";
-}
-
-/// <summary>
-/// Define a validation error contaner
-/// </summary>
 
 [Serializable, DebuggerDisplay("{DebuggerDisplay(),nq}")]
 public class FlatValidationResult
@@ -25,11 +13,17 @@ public class FlatValidationResult
     #region Members
 
     private List<FlatValidationError> errors = new();
+    private List<FlatValidationWarning> warnings = new();
 
     /// <summary>
     /// A collection of errors
     /// </summary>
     public IReadOnlyList<FlatValidationError> Errors => errors;
+
+    /// <summary>
+    /// A collection of warnings
+    /// </summary>
+    public IReadOnlyList<FlatValidationWarning> Warnings => warnings;
 
     /// <summary>
     /// Exception if it occured during validation
@@ -54,14 +48,19 @@ public class FlatValidationResult
     /// <summary>
     /// Creates a new ValidationResult from a collection of errors
     /// </summary>
-    /// <param name="exception">Instance of <see cref="Exception"/> which is later available through the <see cref="Errors"/> property.</param>
+    /// <param name="exception">Instance of <see cref="Exception"/> which is later available through the <see cref="Exception"/> property.</param>
     public FlatValidationResult(Exception exception) => Exception = exception;
 
     /// <summary>
     /// Creates a new ValidationResult from a collection of errors
     /// </summary>
-    /// <param name="errors">Collection of <see cref="FlatValidationError"/> instances which is later available through the <see cref="Errors"/> property.</param>
-    public FlatValidationResult(IEnumerable<FlatValidationError> errors) => AddErrorRange(errors);
+    /// <param name="result">Another instance of the <see cref="FlatValidationResult"/> that has to be cloned.</param>
+    public FlatValidationResult(FlatValidationResult result)
+    {
+        errors.AddRange(result.errors);
+        warnings.AddRange(result.warnings);
+        Exception = result.Exception;
+    }
 
     #endregion // Constructors
 
@@ -76,23 +75,30 @@ public class FlatValidationResult
     #region Public methods
 
     /// <summary>
-    /// Converts the ValidationResult's errors collection into a simple dictionary representation  grouped by PropertyName.
+    /// Converts the ValidationResult's error collection into a simple dictionary representation  grouped by PropertyName.
     /// </summary>
-    /// <returns>A dictionary after grouping by PropertyName.</returns>
+    /// <returns>A dictionary that's grouped by PropertyName.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public Dictionary<string, string[]> ToDictionary() => this.errors.GroupedByPropertyName();
 
     /// <summary>
-    /// Adds an validation error object to the end of the Errors list.
+    /// Converts the ValidationResult's warning collection into a simple dictionary representation  grouped by PropertyName.
+    /// </summary>
+    /// <returns>A dictionary after grouping by PropertyName.</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public Dictionary<string, string[]> WarningToDictionary() => this.warnings.GroupedByPropertyName();
+
+    /// <summary>
+    /// Adds an validation error object to the end of the <see cref="FlatValidationResult.Errors"/> list.
     /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void AddError(FlatValidationError error) => this.errors.Add(error);
 
     /// <summary>
-    /// Adds a collection of validation error objects to the end of the Errors list.
+    /// Adds an validation warning object to the end of the <see cref="FlatValidationResult.Warnings"/> list.
     /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public void AddErrorRange(IEnumerable<FlatValidationError> errors) => this.errors.AddRange(errors);
+    public void AddWarning(FlatValidationWarning warning) => this.warnings.Add(warning);
 
     /// <summary>
     /// Generates a string representation of the error messages separated by new lines.