+++ src/FlatValidator/RuleList.cs
@@ -17,6 +17,9 @@ internal enum RuleType : int
 
     ErrorConst,
 
+    WarningSynch,
+    WarningAsync,
+
     GroupSynch,
     GroupAsync,
     GroupCancelledAsync,
@@ -29,8 +32,8 @@ public struct Rule
     internal Delegate Conditions;
     internal Delegate GroupThen;
     internal Delegate GroupElse;
-    internal Delegate Error;
-    internal string ErrorMessage;
+    internal Delegate FuncMessage;
+    internal string ConstMessage;
     internal Expression MemberSelector1;
     internal Expression MemberSelector2;
     internal Expression MemberSelector3;
@@ -109,8 +112,8 @@ internal struct RuleList<TModel>
         rule.Conditions = conditions;
         rule.GroupThen = groupThen;
         rule.GroupElse = groupElse;
-        rule.Error = error;
-        rule.ErrorMessage = errorMessage;
+        rule.FuncMessage = error;
+        rule.ConstMessage = errorMessage;
         rule.MemberSelector1 = memberSelector1;
         rule.MemberSelector2 = memberSelector2;
         rule.MemberSelector3 = memberSelector3;